feat(panelPilot): adjust panel pilot
